AI-Agent-Project/LangGraph-Tutorial-AI/requirements.txt
# LangGraph Tutorial 项目依赖
# Python 版本要求: >= 3.10

# LangGraph 核心
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0

# LangChain 核心
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# 向量库
chromadb>=0.4.0
langchain-chroma>=0.1.0

# 嵌入模型
openai>=1.0.0
dashscope>=1.14.0

# 工具库
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0

# 其他工具
tenacity>=8.2.0
tiktoken>=0.5.0
rich>=13.0.0
